requirements.txt
fastapi
uvicorn
streamlit
transformers
torch
tensorflow-cpu
sqlalchemy
pymysql